storagedetective-backend/addProduct/requirements.txt
# requirements.txt

# For the web framework
Flask==3.0.3
functions-framework==3.5.0

# For interacting with Google Cloud
google-cloud-aiplatform>=1.55.0
google-cloud-storage==2.16.0
google-cloud-secret-manager>=2.0.0
google-cloud-discoveryengine==0.*
google-cloud-firestore==2.16.0
google-api-python-client>=2.0.0
firebase-admin==6.5.0
google-auth==2.29.0

# For handling images and SCSS
Pillow==10.3.0
libsass==0.22.0

# For environment variables
python-dotenv==1.0.1

# Allow cross-origin requests from your frontend
Flask-Cors==4.0.0

# For making HTTP requests to Figma API
requests==2.31.0
urllib3

#For GitHub
PyGithub==1.*

